+++ packages/editor/break/src/exit-break/queries/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './exitBreakAtEdges';
