+++ packages/editor/break/src/exit-break/transforms/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './exitBreak';
